Refactor layout file.
Define a macro which generates functions to parse and show the keyboard map.
Instead of long lists of manual key indexes use enumerate(). This also allows
the keyboards to be represented in a more natural manner.
Define a function render the keyboard using the new layout format. "\n" is used
to indicate end of row. I am considering using "\t" to define split keyboards.